Fix for estados
