+++ speid/models/__init__.py
@@ -1,7 +1,6 @@
-__all__ = ['Account', 'DocumentBaseMixin', 'Event', 'Request', 'Transaction']
+__all__ = ['Account', 'Event', 'Request', 'Transaction']
 
 from .account import Account
-from .document_base_mixin import DocumentBaseMixin
 from .events import Event
 from .requests import Request
 from .transaction import Transaction
