--- speid/models/document_base_mixin.py
@@ -1,45 +1,0 @@
-from mongoengine import Document, ListField, ReferenceField
-
-from speid.models import Event
-from speid.types import EventType
-
-
-class DocumentBaseMixin(Document):
-    events = ListField(ReferenceField(Event))
-
-    def __init__(self, *args, **values):
-        self.events.append(Event(type=EventType.created))
-        super().__init__(*args, **values)
-
-    def save(
-        self,
-        force_insert=False,
-        validate=True,
-        clean=True,
-        write_concern=None,
-        cascade=None,
-        cascade_kwargs=None,
-        _refs=None,
-        save_condition=None,
-        signal_kwargs=None,
-        **kwargs,
-    ):
-        if len(self.events) > 0:
-            [event.save() for event in self.events]
-        super().save(
-            force_insert,
-            validate,
-            clean,
-            write_concern,
-            cascade,
-            cascade_kwargs,
-            _refs,
-            save_condition,
-            signal_kwargs,
-            **kwargs,
-        )
-
-    def delete(self, signal_kwargs=None, **write_concern):
-        if len(self.events) > 0:
-            [event.delete() for event in self.events]
-        super().delete(signal_kwargs, **write_concern)
